chore: Update dependencies and configuration files
