+++ .prettierignore
@@ -1,5 +1,6 @@
 build
 coverage
+dist
 node_modules
 .eslintrc
 .eslintrc.js
