+++ package.json
@@ -6,6 +6,9 @@
   "license": "MIT",
   "description": "A simple ledger app for tracking your expenses.",
   "entry": "src/main.jsx",
+  "engines": {
+    "node": "18.x || 20.x"
+  },
   "scripts": {
     "dev": "vite --port 8080",
     "start": "vite --port 8080",
@@ -43,7 +46,7 @@
     "eslint": "^8.52.0",
     "eslint-config-prettier": "^9.0.0",
     "eslint-import-resolver-alias": "^1.1.2",
-    "eslint-plugin-import": "^2.29.0",
+    "eslint-plugin-import": "^2.29.1",
     "eslint-plugin-jsx-a11y": "^6.7.1",
     "eslint-plugin-prettier": "^5.0.1",
     "eslint-plugin-react": "^7.33.2",
