+++ src/store/index.js
@@ -1,9 +1,10 @@
+import { validators } from '@luca-financial/luca-schema';
 import { combineReducers, configureStore } from '@reduxjs/toolkit';
 
-import { SchemaKeys, validators } from './lucaSchema';
+import { constants } from './lucaSchema';
 import { createListSlicesFromSchemas } from './schemaDrivenSlice';
 
-const slices = createListSlicesFromSchemas(SchemaKeys, validators);
+const slices = createListSlicesFromSchemas(constants.SchemaKeys, validators);
 
 const localStorageMiddleware = (store) => (next) => (action) => {
   const result = next(action);
@@ -13,7 +14,7 @@ const localStorageMiddleware = (store) => (next) => (action) => {
 
 const rootReducer = combineReducers(
   // iterate over all schema keys and add the reducer for each slice
-  Object.values(SchemaKeys).reduce((acc, schemaKey) => {
+  Object.values(constants.SchemaKeys).reduce((acc, schemaKey) => {
     if (slices[schemaKey]) {
       // extract the reducer from the slice and add it to the accumulator object
       acc[schemaKey] = slices[schemaKey].reducer;
