+++ src/store/lucaSchema/hooks/index.js
@@ -1,4 +1,5 @@
 import useEntities from './useEntities';
+import useLucaSchema from './useLucaSchema';
 import useTransactions from './useTransactions';
 
-export { useEntities, useTransactions };
+export { useEntities, useLucaSchema, useTransactions };
