+++ src/store/lucaSchema/hooks/useEntities.js
@@ -1,10 +1,13 @@
 import { slices } from '@s';
-import { SchemaKeys, useSchemaConfig } from '@s/lucaSchema';
+import { constants, useLucaSchema } from '@s/lucaSchema';
 import { useListSlice } from '@s/schemaDrivenSlice';
 
 export default function useEntities() {
-  const { actions, selectors } = useListSlice(slices, SchemaKeys.ENTITY);
-  const config = useSchemaConfig(SchemaKeys.ENTITY);
+  const { actions, selectors } = useListSlice(
+    slices,
+    constants.SchemaKeys.ENTITY
+  );
+  const config = useLucaSchema(constants.SchemaKeys.ENTITY);
 
   const entities = selectors.selectItems;
   const loadedEntities = selectors.selectLoadedItems;
