+++ src/store/lucaSchema/hooks/useTransactions.js
@@ -1,9 +1,12 @@
 import { slices } from '@s';
-import { SchemaKeys } from '@s/lucaSchema';
+import { constants } from '@s/lucaSchema';
 import { useListSlice } from '@s/schemaDrivenSlice';
 
 export default function useTransactions() {
-  const { actions, selectors } = useListSlice(slices, SchemaKeys.TRANSACTION);
+  const { actions, selectors } = useListSlice(
+    slices,
+    constants.SchemaKeys.TRANSACTION
+  );
   const transactions = selectors.selectItems;
   const loadedTransactions = selectors.selectLoadedItems;
